+++ src/main/java/fr/montreuil/iut/kalos_pokemon/modele/Objet.java
@@ -5,7 +5,6 @@ package fr.montreuil.iut.kalos_pokemon.modele;
  * utile pour faire les calculs de distances entre ennemi-tour-projectile
  */
 public interface Objet {
-    int getX();
-
-    int getY();
+    double getX();
+    double getY();
 }
